Optimize GPU memory
Given Python's scoping semantics, tensors will still occupy space on the
GPU, even when outside of a loop. These tensors, if unused, take up
space on the GPU, causing OOM errors to happen while they can be
avoided. By deleting the tensors, more space is able to be allocated on
the GPUs, allowing a larger number of concurrent processes to run.

Signed-Off-By: Robert Clark <robdclark@outlook.com>