Update reward algorithm to encourage speed
The previous reward algorithm encouraged Mario to both move forward and
kill enemies and grab coins along the way to achieve the highest score.
My desire is to have Mario finish the level as fast as possible, and he
should therefore only be rewarded based on how quickly he finishes a
level, and not have various score-related actions be factored into the
reward.

Signed-Off-By: Robert Clark <robdclark@outlook.com>