Only save unique models
The previous implementation saved models if they weren't equal to the
previous best, but this is inefficient for two reasons:
  1. If a pass through the stage is unique, but managed to grab the same
high score as the previous best, this new model wouldn't be saved.
  2. In phases where models could experienced limited updates during
training, the model used in the testing thread could go through several
iterations without updating, and multiple copies of the same model and
recording would be saved, wasting space.

The new method will only save models that are truly unique by comparing
the list of previously saved models with the latest version.

Signed-Off-By: Robert Clark <robdclark@outlook.com>